refactor: extract TunDevice::send_outbound
