+++ boltconn/src/network/mod.rs
@@ -3,3 +3,4 @@ pub mod dns;
 pub mod egress;
 pub mod packet;
 pub mod tun_device;
+mod unix_tun;
